requirements.txt
Django==1.7.1
Pillow==2.6.1
djangorestframework==2.4.4
psycopg2==2.5.4
wsgiref==0.1.2
